+++ tests/tests.rs
@@ -227,7 +227,7 @@ pub fn convert_to_png() {
         let min_toned = tone(*sorted.first().unwrap());
 
         for (x, y, pixel) in png_buffer.enumerate_pixels_mut() {
-            let v = data[(y * size.0 as u32 + x) as usize];
+            let v = data[(y as usize * size.0 + x as usize)];
             let v = (v - min) / (max - min);
             let v = tone(v);
